[dev] MethodUserGetStrategiesStatistics WIP & bugfixes

+++ db/api.sql
@@ -712,7 +712,7 @@ CREATE OR REPLACE FUNCTION api.fun_user_list_followed_experts(a_user_id bigint)
 RETURNS table (
     "expert_id" bigint,
     "name" varchar,
-    "follower_count" int,
+    "follower_count" bigint,
     "description" varchar,
     "social_media" varchar,
     "risk_score" double precision,
